chore: rename types to felt
