+++ crates/committer/src/deserialization/cast.rs
@@ -2,12 +2,12 @@ use crate::deserialization::types::{ContractAddress, ContractState};
 use crate::deserialization::types::{
     Input, RawInput, StarknetStorageKey, StarknetStorageValue, StateDiff,
 };
+use crate::felt::Felt;
 use crate::hash::hash_trait::HashOutput;
 use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
 use crate::patricia_merkle_tree::types::TreeHeight;
 use crate::storage::errors::DeserializationError;
 use crate::storage::storage_trait::{StorageKey, StorageValue};
-use crate::types::Felt;
 
 use std::collections::HashMap;
