+++ crates/committer/src/deserialization/read_test.rs
@@ -4,12 +4,12 @@ use std::collections::HashMap;
 use crate::deserialization::types::{
     ContractAddress, ContractState, Input, StarknetStorageKey, StarknetStorageValue, StateDiff,
 };
+use crate::felt::Felt;
 use crate::hash::hash_trait::HashOutput;
 use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
 use crate::patricia_merkle_tree::types::TreeHeight;
 use crate::storage::errors::DeserializationError;
 use crate::storage::storage_trait::{StorageKey, StorageValue};
-use crate::types::Felt;
 
 use super::parse_input;
