+++ crates/committer/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod deserialization;
+pub mod felt;
 pub mod hash;
 pub mod patricia_merkle_tree;
 pub mod storage;
-pub mod types;
