+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/hash_function.rs
@@ -1,11 +1,11 @@
 use std::marker::PhantomData;
 
+use crate::felt::Felt;
 use crate::hash::hash_trait::{HashFunction, HashInputPair, HashOutput};
 use crate::patricia_merkle_tree::node_data::inner_node::{
     BinaryData, EdgeData, NodeData, PathToBottom,
 };
 use crate::patricia_merkle_tree::node_data::leaf::{LeafData, LeafDataTrait};
-use crate::types::Felt;
 
 #[cfg(test)]
 #[path = "hash_function_test.rs"]
