+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/tree_test.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use crate::felt::Felt;
 use crate::hash::hash_trait::HashOutput;
 use crate::hash::pedersen::PedersenHashFunction;
 use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, FilledNode};
@@ -14,7 +15,6 @@ use crate::patricia_merkle_tree::updated_skeleton_tree::node::UpdatedSkeletonNod
 use crate::patricia_merkle_tree::updated_skeleton_tree::tree::{
     UpdatedSkeletonTree, UpdatedSkeletonTreeImpl,
 };
-use crate::types::Felt;
 
 #[tokio::test(flavor = "multi_thread")]
 /// This test is a sanity test for computing the root hash of the patricia merkle tree with a single node that is a leaf with hash==1.
